treewide: format all inactive Nix files
Inactive means that there are no open PRs with activity in the
last month that touch those files.
A bunch later, we can do another pass to get the rest.

Doing it this way makes ensures that we don't cause any conflicts for
recently active PRs that would be ready to merge.
Only once those PRs get updated, CI will kick in and require the files
to be formatted.

Furthermore, this makes sure that we can merge this PR without having
to constantly rebase it!

This can be verified using

    nix-build https://gist.github.com/infinisil/4b7a1a1e5db681d04446e73e39048aec/archive/5439e16eda66856aed747db3b45ed595835de7c6.tar.gz \
      --argstr baseRev 5180ff0970d05347aa87492f64ed84a6e1e24b33
    result/bin/apply-formatting $NIXPKGS_PATH

+++ pkgs/os-specific/linux/atop/default.nix
@@ -1,15 +1,16 @@
-{ lib
-, stdenv
-, fetchurl
-, glib
-, zlib
-, ncurses
-, pkg-config
-, findutils
-, systemd
-, python3
-# makes the package unfree via pynvml
-, withAtopgpu ? false
+{
+  lib,
+  stdenv,
+  fetchurl,
+  glib,
+  zlib,
+  ncurses,
+  pkg-config,
+  findutils,
+  systemd,
+  python3,
+  # makes the package unfree via pynvml
+  withAtopgpu ? false,
 }:
 
 stdenv.mkDerivation rec {
@@ -21,22 +22,16 @@ stdenv.mkDerivation rec {
     hash = "sha256-56ZzzyyCV4592C7LDeyD/Z7LMIKLJWHCip+lqvddX5M=";
   };
 
-  nativeBuildInputs = lib.optionals withAtopgpu [
-    python3.pkgs.wrapPython
-  ];
+  nativeBuildInputs = lib.optionals withAtopgpu [ python3.pkgs.wrapPython ];
 
   buildInputs = [
     glib
     zlib
     ncurses
     pkg-config
-  ] ++ lib.optionals withAtopgpu [
-    python3
-  ];
+  ] ++ lib.optionals withAtopgpu [ python3 ];
 
-  pythonPath = lib.optionals withAtopgpu [
-    python3.pkgs.pynvml
-  ];
+  pythonPath = lib.optionals withAtopgpu [ python3.pkgs.pynvml ];
 
   makeFlags = [
     "DESTDIR=$(out)"
@@ -70,14 +65,21 @@ stdenv.mkDerivation rec {
     mkdir -p $out/bin
   '';
 
-  postInstall = ''
-    # Remove extra files we don't need
-    rm -r $out/{var,etc} $out/bin/atop{sar,}-${version}
-  '' + (if withAtopgpu then ''
-    wrapPythonPrograms
-  '' else ''
-    rm $out/lib/systemd/system/atopgpu.service $out/bin/atopgpud $out/share/man/man8/atopgpud.8
-  '');
+  postInstall =
+    ''
+      # Remove extra files we don't need
+      rm -r $out/{var,etc} $out/bin/atop{sar,}-${version}
+    ''
+    + (
+      if withAtopgpu then
+        ''
+          wrapPythonPrograms
+        ''
+      else
+        ''
+          rm $out/lib/systemd/system/atopgpu.service $out/bin/atopgpud $out/share/man/man8/atopgpud.8
+        ''
+    );
 
   meta = with lib; {
     platforms = platforms.linux;
